skateboarding/ig_bot/requirements.txt
praw==6.0.0
psycopg2-binary==2.7.6.1
